Transfer DRP pipelines from other packages and clean them up.
This commit includes transfers from pipe_tasks, obs_subaru, obs_lsst,
and obs_decam.

This transfer was originally performed over a period of several weeks,
with some cleanups and reorganization done between those code
transfers, as well as changes to embedded paths to reflect new
locations.  This commit squashes all of that together, which is not
ideal, but the the original history is simply confusing, because
transfers done on top of cleanups inevitably don't look like pure
transfers anymore, and hence it's impossible to map most commits to
previous states of the origin git repositorie.

Cleanups include comment and description clarifications, a lot of
whitespace (to satisfy yamllint), and a few true/false capitalization
issues.  It converts all multi-line strings to use '|' (preserve
newlines) instead of `>` (delete newlines).  We'd ideally want
something that translates two newlines into one (as Markdown and
reStructuredText do), but if we preserve them all we can at least do
that later.