Add pipeline definition for pipelines_check.
The pipelines_check scripts originally ran a subset of the pipe_tasks
generic DRP pipeline, despite running on HSC data.  This replacement
instead pulls out a subset of the HSC DRP pipeline (though some
refactoring is still needed to move that to ingredients/), and by
making it a separate pipeline instead of a command-line subset we both
document what's in pipelines_check and make it easier to run against
other data repositories.