Refactor pipeline definitions.
This mostly just moves HSC content from the "pipelines" directory to
"ingredients", bringing HSC more in line with the precedent set by
other instruments.  There's certainly more to do to clean up the
organization of these (that's DM-34039), but this is enough to set the
stage for new ingredients files to be included by the same pipelines.

This also establishes a new naming convention for the files in
"ingredients" (well, "convention" might be overstating it a bit,
especially as these new names aren't likely to survive DM-34039),
reflecting the fact that these are not intended to have the same
scope as the full-DRP pipelines in the "pipelines" directory.

Finally, I've tweaked the docstrings for various pipelines a bit to
improve consistency and try to make it clearer what kinds of changes
should go where.