Reorganize pipelines import structure and fix configs.
This addresses a lot of "backwards" imports, in which the very-much
concrete and complete HSC/DRP-RC2 pipeline was imported by other
pipelines, by moving its contents to new pipelines in ingredients.
This also reduces duplication between HSC and DECam.

While it's mostly a refactor, it does change the expanded pipelines in
a few ways:

- HSC/DRP-Prod now has the correct (full-SSP) configuration for the
  FGCM tasks, mostly by avoiding RC2-specific overrides.

- Two ill-defined subsets (singleFrame and multiVisit) have been
  removed from all pipelines.

- Superfluous "step" subsets have been removed (including all steps in
  the ci_hsc pipelines).

- faro tasks have been removed from HSC/DRP-Prod and ci_hsc.