Add UpdateVisitSummaryTask to DRP pipelines.
Changes in pipe_tasks will take care of making existing tasks use this
new task's outputs instead of the old visit summaries, so most of the
changes here are about redefining subsets to be compatible with the
new overall flow.

Many of the DECam changes are actually removals of Gen2 config options
that now do nothing (but previously did the equivalent of Gen3
connection changes).