Remove unnecessary tasks from fakes pipeline.
This pipeline is soon to be deprecated anyway, but since we're still
nominally maintaining it it's useful to drop these tasks that don't
serve a purpose there: they're all things that really only need to be
run once, before the fakes are inserted, and rerunning them after the
fakes have been inserted either doesn't change them or adds circular
measurements (e.g. PSF summary stats on fakes inserted with the PSF
model we're trying to summaries) that we don't.