Remove step8 and move its analysis tasks upstream
Now that we're running RC2/DC2s with the sasquatch butler for all
steps, we can move the analysis tasks to the step which creates
their input data products.

Move source table consolidation to the global step2 stage, and
add consolidateSourceTable to the per-visit sourceTable cluster
for [writeRecalibrated|transform|consolidate]SourceTable.