Refactor ci_imsim to use reprocessVisitImage
ci_imsim doesn't have much of the visit-level processing (e.g. fgcmcal), but
we still have to run reprocessVisitImage to get the full source catalog, since
calibrateImage only produces a catalog of stars, with fewer measurements.
This then requires treating the stars catalog as a "preSourceTable" and adding
some other tasks that would normally only be run in a "full" pipeline.