Use reprocessVisitImage output in forced phot and diffim.
For forced photometry on direct images, this should cause us to start
using the skyCorr background (via its subtraction from pvi) where we
were not before - I suspect due to an oversight.  We no longer need to
read the visit summary to get the latest versions of everything else,
since those are also in pvi.

For diffim, this change means that we're using the final WCS instead
of the preliminary one (addressing a TODO comment for DM-31292 in
GetTemplateTask) when warping the template to match the science image,
and we're doing the subtraction after applying the skyCorr background
(when it is available), which probably doesn't matter in this context
but seems good for consistency.  This also gets the final WCS and
PhotoCalib into detectAndMeasure and hence into the Local* measurements
that are used to calibrate.  That's something we could have been doing
in the past if we'd ever set doApplyExternalCalibration=True in
subtractImages in DRP, but we were not.