Switch to CalibrateImage in ComCam and reconfigure quickLooks.
The task-default configuration of CalibrateImage is already tuned (due
to its use in AP) for performance rather than completeness, so we can
use that in quickLook instead of explicitly undoing the many
CalibrateTask-emulating config overrides we have in the base
ingredients files.  We do want the config overrides that rename the
output connections (at least until DM-47320), however, so we have to
repeat them in the quickLook pipelines.