Drop now-unnecessary analyzePreSourceTable configurations.
This configuration is now part of visitQualityCore.yaml in
analysis_tools.  By not excluding it from other pipelines here it will
be explicitly added to them, but since most of those pipelines are
really only defined by what's in their steps, only ci_hsc will actually
be affected, and that's good, because it's the only test coverage we
can easily get for this configuration.