Use v2's recalibrated_star_unstandardized in v1.
This avoids perpetuating the conflict on the definition of the
'source' dataset, though we will need to migrate data repositories
eventually as well.  While the new name doesn't fit in well with the
rest of the v1 names, it seems better to just use the v2 name than to
clutter repos with yet another dataset type name, especially given
that we're (slowly) migrating away from v1 entirely.