Stand-up LATISS ingredients pipeline YAML
On DM-49600 it was discovered that the obs_lsst config overrides for the
calibrateImage task in LATISS were not being picked up correctly due to
a subsequent config override in `drp_pipe/config/calibrateImage.py`.
Standing up a LATISS ingredients YAML and setting the appropriate config
override here resolves this issue, as this file is last in the chain.
The offending config has been removed from obs_lsst to prevent possible
future confusion as to it being set there.