Add splitPrimaryObjectForcedSource to DRP pipelines.
As with the Object split, this involve renaming the inputs with _all
in v2 and renaming the outputs with _primary in v1.

This applied only to the Object forced photometry because non-primary
DIAObjects are stripped out earlier.

Because we're now regarding tract-level forced consolidation as
undesirable, we're running SplitPrimaryTask on the patch-level
outputs only.  I have not removed the tract-level consolidations from
the v1 tasks (they are already gone from v2).